move testcase from src to tests
