+++ tests/data/testcases/oscar_testcases.py
